backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# HTTP client
httpx==0.25.2

# Image processing (lightweight)
pillow==10.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Basic utilities
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Lightweight ML alternatives (commented out for demo)
# numpy==1.24.4
# pandas==2.1.4

# Heavy ML libraries removed for demo (uncomment for production):
# openai==1.3.7
# transformers==4.36.0
# torch==2.1.1
# torchvision==0.16.1
# scikit-learn==1.3.2
# geopy==2.4.1
# folium==0.15.0
# redis==5.0.1
# celery==5.3.4
